update cards add upgrade cancel

+++ database/migrations/0002_01_01_000006_create_buildings_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->foreignId('details_id')
                   ->constrained('building_details')
                   ->onDelete('cascade');
-            $table->bigInteger('effect_value')->nullable();
+            $table->decimal('effect_value', 8, 2)->nullable();
             $table->integer('level')->default(1);
             $table->integer('build_time')->nullable();
             $table->timestamps();
